refactor(mcp): relocate the v2 gateway to litellm/proxy/gateway/mcp
Moves the clean-room gateway package out of litellm/proxy/_experimental/mcp_server/v2
into its final home litellm/proxy/gateway/mcp (the folder graduation the migration plan
had slated for Phase 2, pulled forward so all further work lands at the final path with
no later import churn). Sitting outside _experimental/mcp_server also sharpens the
clean-room boundary: the gateway no longer nests under the v1 MCP code it must not import.

Pure relocation, no behavior change: git-tracked renames preserve history; relative imports
are unaffected; the test's absolute imports, the README's basedpyright command, and the
pyrightconfig venvPath depth are updated for the new location. Tests move to
tests/mcp_tests/gateway. black, ruff, basedpyright strict, and the 16 resolver tests all
pass at the new path.
