+++ litellm/proxy/gateway/mcp/oauth/httpx_auth.py
