+++ litellm/proxy/gateway/mcp/pyrightconfig.json
@@ -2,7 +2,7 @@
     "include": ["."],
     "typeCheckingMode": "strict",
     "pythonVersion": "3.10",
-    "venvPath": "../../../../..",
+    "venvPath": "../../../..",
     "venv": ".venv",
     "reportMissingTypeStubs": false,
     "reportUnknownMemberType": true,
